PhysiCellXMLRules

Documentation for PhysiCellXMLRules.

Convert PhysiCell rules between CSV and XML formats. Enables conversion between https://github.com/MathCancer/PhysiCell/ and https://github.com/drbergman/PhysiCell/.

PhysiCellXMLRules.exportRulesToCSV

PhysiCellXMLRules.standardizeCustomName

PhysiCellXMLRules.standardizeCustomNameExport

PhysiCellXMLRules.writeRules

PhysiCellXMLRules.exportRulesToCSV — Method

exportRulesToCSV(path_to_csv::AbstractString, path_to_xml::AbstractString)

Export the rules from the XML file at path_to_xml to the CSV file at path_to_csv.

Note: this is not the inverse of writeRules as writeRules discards comments in the original CSV and exportRulesToCSV adds comments to the exported CSV file.

PhysiCellXMLRules.standardizeCustomName — Method

standardizeCustomName(name::AbstractString)

If the name for a signal or behavior starts with "custom:", use the synonym "custom <name>" instead.

pcvct uses : to use attributes to specify elements in an XML path. So, we use custom <name> to avoid incorrect splitting on :.

PhysiCellXMLRules.standardizeCustomNameExport — Method

standardizeCustomNameExport(name::AbstractString)

If the name for a signal or behavior starts with "custom ", use the synonym "custom:<name>" instead when exporting to a CSV.

Both are acceptable, but this function will convert it to the more standard format in PhysiCell.

PhysiCellXMLRules.writeRules — Method

writeRules(path_to_xml::AbstractString, path_to_csv::AbstractString)

Write the rules from the CSV file at path_to_csv to the XML file at path_to_xml.

Note: this is not the inverse of exportRulesToCSV as writeRules discards comments in the original CSV and exportRulesToCSV adds comments to the exported CSV file.