Write XML

Writes the XML file for behavior rulesets. Note, when importing a CSV to write to XML, the full expressiveness of the XML format is not available. In particular, the following cannot be read in from a CSV:

hierarchical rulesets (those with mediators and aggregators below the top two levels)

mediator/aggregator functions

attenuator/accumulator behaviors (only setters)

PhysiCellXMLRules.standardizeCustomName — Method

standardizeCustomName(name::AbstractString)

If the name for a signal or behavior starts with "custom:", use the synonym "custom <name>" instead.

pcvct uses : to use attributes to specify elements in an XML path. So, we use custom <name> to avoid incorrect splitting on :.

PhysiCellXMLRules.writeXMLRules — Method

writeXMLRules(path_to_xml::AbstractString, path_to_csv::AbstractString; force::Bool=false)

Write the rules from the CSV file at path_to_csv to the XML file at path_to_xml.

If force is set to true, the function will overwrite the existing XML file at path_to_xml if it exists. Note: this is not the inverse of exportCSVRules as writeXMLRules discards comments in the original CSV and exportCSVRules adds comments to the exported CSV file.